+++ sphaera/diffeq/_impl/misc.py
@@ -163,7 +163,7 @@ def _optimal_step_size(last_step, mean_error_ratio, safety=0.9, ifactor=10.0, df
     if mean_error_ratio == 0:
         return last_step * ifactor
     if mean_error_ratio < 1:
-        dfactor = _convert_to_tensor(1, dtype=torch.float64, device=mean_error_ratio.device)
+        dfactor = _convert_to_tensor(1, dtype=torch.float32, device=mean_error_ratio.device)
     error_ratio = torch.sqrt(mean_error_ratio).to(last_step)
     exponent = torch.tensor(1 / order).to(last_step)
     factor = torch.max(1 / ifactor, torch.min(error_ratio**exponent / safety, 1 / dfactor))
